implemented QueryFormatter
