+++ package.json
@@ -123,6 +123,7 @@
     "debug": "^4.1.1",
     "dotenv": "^8.2.0",
     "glob": "^7.1.6",
+    "hdb-pool": "^0.1.6",
     "js-yaml": "^3.14.0",
     "mkdirp": "^1.0.4",
     "reflect-metadata": "^0.1.13",
