+++ src/connection/Connection.ts
@@ -40,6 +40,7 @@ import {IsolationLevel} from "../driver/types/IsolationLevel";
 import {AuroraDataApiDriver} from "../driver/aurora-data-api/AuroraDataApiDriver";
 import {DriverUtils} from "../driver/DriverUtils";
 import {ReplicationMode} from "../driver/types/ReplicationMode";
+import { QueryFormatter } from "../query-builder/QueryFormatter";
 
 /**
  * Connection is a single database ORM connection to a specific database.
@@ -435,6 +436,14 @@ export class Connection {
         } else {
             return new SelectQueryBuilder(this, entityOrRunner as QueryRunner|undefined);
         }
+    }
+
+
+    createQueryFormatter(): QueryFormatter {
+        if (this instanceof MongoEntityManager)
+            throw new Error(`Query Formatter is not supported by MongoDB.`);
+
+        return new QueryFormatter(this);
     }
 
     /**
