+++ src/query-builder/QueryBuilder.ts
@@ -125,14 +125,6 @@ export abstract class QueryBuilder<Entity> {
     // -------------------------------------------------------------------------
 
     /**
-     * Set full query
-     */
-    setQuery(sql: string): this {
-        this.expressionMap.primarySql = sql;
-        return this;
-    }
-
-    /**
      * Creates SELECT query.
      * Replaces all previous selections if they exist.
      */
@@ -370,26 +362,6 @@ export abstract class QueryBuilder<Entity> {
         return this;
     }
 
-    addAlias(entityTarget: EntityTarget<any>, aliasName: string): this {
-        if (this.connection.hasMetadata(entityTarget)) {
-            const metadata = this.connection.getMetadata(entityTarget);
-
-            this.expressionMap.createAlias({
-                type: "other",
-                name: aliasName,
-                metadata: this.connection.getMetadata(entityTarget),
-                tablePath: metadata.tablePath,
-            });
-        }
-
-        return this;
-    }
-
-    setAliases(aliases: { [aliasName: string]: EntityTarget<any> }): this {
-        Object.keys(aliases).forEach((key) => this.addAlias(aliases[key], key));
-        return this;
-    }
-
     /**
      * Gets all parameters.
      */
@@ -600,65 +572,13 @@ export abstract class QueryBuilder<Entity> {
         }
     }
 
-
-    /**
-     * Escape RegExp
-     */
-
-    protected escapeRegExp(s: String) {
-        return s.replace(/[.*+\-?^${}()|[\]\\]/g, "\\$&");
-    }
-
-    /**
-     * Replaces all entity's aliases to name in the given statement.
-     */
-
-    protected replacePropertyAliases(statement: string) {
-        const replacements: { [key: string]: string } = {};
-
-        for (const alias of this.expressionMap.aliases) {
-            if (alias.tablePath && alias.type === "other") {
-                replacements[alias.name] =
-                    this.getTableName(alias.tablePath) +
-                    " " +
-                    this.escape(alias.name);
-                    
-                const columnAliases: string[] = [];
-
-                for (const column of alias.metadata.columns) {
-                    const columnAlias = 
-                        `${alias.name}.${column.propertyName}`  +
-                        " AS " +
-                        this.escape(column.propertyName);
-
-                    replacements[`${alias.name}.${column.propertyName}`] = columnAlias;
-                    columnAliases.push(columnAlias);
-                }
-        
-                replacements[`${alias.name}.*`] = columnAliases.join(", ");
-            }
-        }
-        
-        const replacementKeys = Object.keys(replacements);
-
-        statement = statement.replace(
-            new RegExp(
-               `(?<=[ =\(]|^)` + `\\&(${replacementKeys.map(this.escapeRegExp).join("|")})` + `(?=[ =\)\,]|$)`,
-                "gm"
-            ),
-            (_, p) => replacements[p]
-        );
-
-        return statement;
-    }
-
     /**
      * Replaces all entity's propertyName to name in the given statement.
      */
     protected replacePropertyNames(statement: string) {
         // Escape special characters in regular expressions
         // Per https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_Expressions#Escaping
-        
+        const escapeRegExp = (s: String) => s.replace(/[.*+\-?^${}()|[\]\\]/g, "\\$&");
 
         for (const alias of this.expressionMap.aliases) {
             if (!alias.hasMetadata) continue;
@@ -705,7 +625,7 @@ export abstract class QueryBuilder<Entity> {
             if (replacementKeys.length) {
                 statement = statement.replace(new RegExp(
                     `(?<=[ =\(]|^.{0})` +
-                    `${this.escapeRegExp(replaceAliasNamePrefix)}(${replacementKeys.map(this.escapeRegExp).join("|")})` +
+                    `${escapeRegExp(replaceAliasNamePrefix)}(${replacementKeys.map(escapeRegExp).join("|")})` +
                     `(?=[ =\)\,]|.{0}$)`,
                     "gm"
                 ), (_, p) =>
@@ -715,13 +635,6 @@ export abstract class QueryBuilder<Entity> {
         }
 
         return statement;
-    }
-
-    protected createPrimarySql(): string {
-        let sql = this.expressionMap.primarySql;
-        sql = this.replacePropertyAliases(sql);
-        sql = this.replacePropertyNames(sql);
-        return sql;
     }
 
     protected createComment(): string {
