+++ src/query-builder/QueryExpressionMap.ts
@@ -276,11 +276,6 @@ export class QueryExpressionMap {
      */
     comment?: string;
 
-    /**
-     * Full sql query
-     */
-    primarySql: string = "";
-
     // -------------------------------------------------------------------------
     // Constructor
     // -------------------------------------------------------------------------
